backend/requirements.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1

# Graph Database
neo4j==5.18.0

# AI & Orchestration (LangChain & LangGraph)
langchain==0.1.13
langchain-openai==0.1.1
langchain-core==0.1.33
langchain-community==0.0.29
langgraph==0.0.30
langsmith==0.1.31

# Vector DB & Semantic Caching
redis==5.0.3

# Data Validation & CLI
pydantic==2.6.4
rich==13.7.1